// src/graphql/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type AnnotationConnection = Connection & {
  __typename?: 'AnnotationConnection';
  count: Scalars['Int']['output'];
  edges?: Maybe<Array<Maybe<AnnotationEdge>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type Build = Node & {
  __typename?: 'Build';
  annotations?: Maybe<AnnotationConnection>;
  /** The current blocked state of the build */
  blockedState?: Maybe<BuildBlockedStates>;
  /** The branch for the build */
  branch: Scalars['String']['output'];
  /** The time when the build was cancelled */
  canceledAt?: Maybe<Scalars['DateTime']['output']>;
  /** The user who canceled this build. If the build was canceled, and this value is null, then it was canceled automatically by Buildkite */
  canceledBy?: Maybe<User>;
  /** The fully-qualified commit for the build */
  commit: Scalars['String']['output'];
  /** The time when the build was created */
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  createdBy?: Maybe<BuildCreator>;
  /** Custom environment variables passed to this build */
  env?: Maybe<Array<Scalars['String']['output']>>;
  /** The time when the build finished */
  finishedAt?: Maybe<Scalars['DateTime']['output']>;
  id: Scalars['ID']['output'];
  jobs?: Maybe<JobConnection>;
  /** The message for the build */
  message?: Maybe<Scalars['String']['output']>;
  metaData?: Maybe<BuildMetaDataConnection>;
  /** The number of the build */
  number: Scalars['Int']['output'];
  organization: Organization;
  pipeline: Pipeline;
  pullRequest?: Maybe<PullRequest>;
  /** The build that this build was rebuilt from */
  rebuiltFrom?: Maybe<Build>;
  /** The time when the build became scheduled for running */
  scheduledAt?: Maybe<Scalars['DateTime']['output']>;
  /** Where the build was created */
  source: BuildSource;
  /** The time when the build started running */
  startedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The current state of the build */
  state: BuildStates;
  /** The job that this build was triggered from */
  triggeredFrom?: Maybe<JobTypeTrigger>;
  /** The URL for the build */
  url: Scalars['String']['output'];
  /** The UUID for the build */
  uuid: Scalars['String']['output'];
};
// ...
export enum BuildBlockedStates {
  /** The blocked build is failed */
  Failed = 'FAILED',
  /** The blocked build is passed */
  Passed = 'PASSED',
  /** The blocked build is running */
  Running = 'RUNNING'
}
// ...
export type BuildCreator = UnregisteredUser | User;

export type BuildEdge = {
  __typename?: 'BuildEdge';
  cursor: Scalars['String']['output'];
  node?: Maybe<Build>;
};

/** A comment on a build */
export type BuildMetaData = {
  __typename?: 'BuildMetaData';
  /** The key used to set this meta data */
  key: Scalars['String']['output'];
  /** The value set to this meta data */
  value: Scalars['String']['output'];
};

export type BuildMetaDataConnection = Connection & {
  __typename?: 'BuildMetaDataConnection';
  count: Scalars['Int']['output'];
  edges?: Maybe<Array<Maybe<BuildMetaDataEdge>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export enum BuildStates {
  /** The build is blocked */
  Blocked = 'BLOCKED',
  /** The build was canceled */
  Canceled = 'CANCELED',
  /** The build is currently being canceled */
  Canceling = 'CANCELING',
  /** The build is currently being created */
  Creating = 'CREATING',
  /** The build failed */
  Failed = 'FAILED',
  /** The build is failing */
  Failing = 'FAILING',
  /** The build wasn't run */
  NotRun = 'NOT_RUN',
  /** The build passed */
  Passed = 'PASSED',
  /** The build is currently running jobs */
  Running = 'RUNNING',
  /** The build has yet to start running jobs */
  Scheduled = 'SCHEDULED',
  /** The build was skipped */
  Skipped = 'SKIPPED'
}
// ...
export type Job = JobTypeBlock | JobTypeCommand | JobTypeTrigger | JobTypeWait;
// ...
export type JobConnection = Connection & {
  __typename?: 'JobConnection';
  count: Scalars['Int']['output'];
  edges?: Maybe<Array<Maybe<JobEdge>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type Node = {
  /** ID of the object. */
  id: Scalars['ID']['output'];
};
// ...
export type Organization = Node & {
  __typename?: 'Organization';
  /** Returns agent access tokens for an Organization. By default returns all tokens, whether revoked or non-revoked. */
  agentTokens?: Maybe<AgentTokenConnection>;
  agents?: Maybe<AgentConnection>;
  /** A space-separated allowlist of IP addresses that can access the organization via the GraphQL or REST API */
  allowedApiIpAddresses?: Maybe<Scalars['String']['output']>;
  /** Returns user API access tokens that can access this organization */
  apiAccessTokens: OrganizationApiAccessTokenConnection;
  auditEvents?: Maybe<OrganizationAuditEventConnection>;
  /** Returns active banners for this organization. */
  banners: OrganizationBannerConnection;
  /** Return cluster in the Organization by UUID */
  cluster?: Maybe<Cluster>;
  /** Returns clusters for an Organization */
  clusters?: Maybe<ClusterConnection>;
  /** The URL to an icon representing this organization */
  iconUrl?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  invitations?: Maybe<OrganizationInvitationConnection>;
  /** Whether teams is enabled for this organization */
  isTeamsEnabled: Scalars['Boolean']['output'];
  jobs?: Maybe<JobConnection>;
  /** Returns users within the organization */
  members?: Maybe<OrganizationMemberConnection>;
  /** Whether this organization requires 2FA to access (Please note that this is a beta feature and is not yet available to all organizations.) */
  membersRequireTwoFactorAuthentication: Scalars['Boolean']['output'];
  /** The name of the organization */
  name: Scalars['String']['output'];
  permissions: OrganizationPermissions;
  /** Return all the pipeline templates the current user has access to for this organization */
  pipelineTemplates?: Maybe<PipelineTemplateConnection>;
  /** Return all the pipelines the current user has access to for this organization */
  pipelines?: Maybe<PipelineConnection>;
  /** Whether this organization is visible to everyone, including people outside it */
  public: Scalars['Boolean']['output'];
  /** Return all the registries the current user has access to for this organization */
  registries?: Maybe<RegistryConnection>;
  /** Returns the repository providers for this organization */
  repositoryProviders?: Maybe<Array<Maybe<OrganizationRepositoryProvider>>>;
  /** API tokens with access to this organization will be automatically revoked after this many seconds of inactivity. A `null` value indicates never revoke inactive tokens. */
  revokeInactiveTokensAfter?: Maybe<RevokeInactiveTokenPeriod>;
  /** Returns rules for an Organization */
  rules?: Maybe<RuleConnection>;
  /** The slug used to represent the organization in URLs */
  slug: Scalars['String']['output'];
  /** The single sign-on configuration of this organization */
  sso?: Maybe<OrganizationSso>;
  /** Single sign on providers created for an organization */
  ssoProviders?: Maybe<SsoProviderConnection>;
  /** Return all the suite the current user has access to for this organization */
  suites?: Maybe<SuiteConnection>;
  /** Returns teams within the organization that the viewer can see */
  teams?: Maybe<TeamConnection>;
  /** Returns the resource usage data for this organization. */
  usage: UsageUnionConnection;
  /** The public UUID for this organization */
  uuid: Scalars['String']['output'];
};
// ...
export type Pipeline = Node & {
  __typename?: 'Pipeline';
  /** Whether existing builds can be rebuilt as new builds. */
  allowRebuilds?: Maybe<Scalars['Boolean']['output']>;
  /** Whether this pipeline has been archived */
  archived: Scalars['Boolean']['output'];
  /** The time when the pipeline was archived */
  archivedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The user that archived this pipeline */
  archivedBy?: Maybe<User>;
  /** The URL for build badge to the show the current build state for a pipeline. */
  badgeURL: Scalars['String']['output'];
  /** A branch filter pattern to limit which pushed branches trigger builds on this pipeline. */
  branchConfiguration?: Maybe<Scalars['String']['output']>;
  /** Returns the builds for this pipeline */
  builds?: Maybe<BuildConnection>;
  /** When a new build is created on a branch, any previous builds that are running on the same branch will be automatically cancelled */
  cancelIntermediateBuilds: Scalars['Boolean']['output'];
  /** Limit which branches build cancelling applies to, for example `!main` will ensure that the main branch won't have it's builds automatically cancelled. */
  cancelIntermediateBuildsBranchFilter?: Maybe<Scalars['String']['output']>;
  cluster?: Maybe<Cluster>;
  /** The color of the pipeline */
  color?: Maybe<Scalars['String']['output']>;
  /** The shortest length to which any git commit ID may be truncated while guaranteeing referring to a unique commit */
  commitShortLength: Scalars['Int']['output'];
  /** The time when the pipeline was created */
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  /** The user who created the pipeline */
  createdBy?: Maybe<User>;
  /** The default branch for this pipeline */
  defaultBranch?: Maybe<Scalars['String']['output']>;
  /** The default timeout in minutes for all command steps in this pipeline. This can still be overridden in any command step */
  defaultTimeoutInMinutes?: Maybe<Scalars['Int']['output']>;
  /** The short description of the pipeline */
  description?: Maybe<Scalars['String']['output']>;
  /** The emoji of the pipeline */
  emoji?: Maybe<Scalars['String']['output']>;
  /** Returns true if the viewer has favorited this pipeline */
  favorite: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  jobs?: Maybe<JobConnection>;
  /** The maximum timeout in minutes for all command steps in this pipeline. Any command step without a timeout or with a timeout greater than this value will be set to this value. */
  maximumTimeoutInMinutes?: Maybe<Scalars['Int']['output']>;
  metrics?: Maybe<PipelineMetricConnection>;
  /** The name of the pipeline */
  name: Scalars['String']['output'];
  /** The next build number */
  nextBuildNumber: Scalars['Int']['output'];
  organization: Organization;
  permissions: PipelinePermissions;
  pipelineTemplate?: Maybe<PipelineTemplate>;
  /**
   * Whether this pipeline is visible to everyone, including people outside this organization
   * @deprecated Use `visibility`
   */
  public: Scalars['Boolean']['output'];
  /** The repository for this pipeline */
  repository?: Maybe<Repository>;
  /** Schedules for this pipeline */
  schedules?: Maybe<PipelineScheduleConnection>;
  /** When a new build is created on a branch, any previous builds that haven't yet started on the same branch will be automatically marked as skipped. */
  skipIntermediateBuilds: Scalars['Boolean']['output'];
  /** Limit which branches build skipping applies to, for example `!main` will ensure that the main branch won't have it's builds automatically skipped. */
  skipIntermediateBuildsBranchFilter?: Maybe<Scalars['String']['output']>;
  /** The slug of the pipeline */
  slug: Scalars['String']['output'];
  steps?: Maybe<PipelineSteps>;
  /** Tags that have been given to this pipeline */
  tags: Array<PipelineTag>;
  /** Teams associated with this pipeline */
  teams?: Maybe<TeamPipelineConnection>;
  /** The URL for the pipeline */
  url: Scalars['String']['output'];
  /** The UUID of the pipeline */
  uuid: Scalars['String']['output'];
  /** Whether this pipeline is visible to everyone, including people outside this organization */
  visibility: PipelineVisibility;
  /** The URL to use in your repository settings for commit webhooks */
  webhookURL: Scalars['String']['output'];
};
// ...
export type User = Node & {
  __typename?: 'User';
  avatar: Avatar;
  /** If this user account is an official bot managed by Buildkite */
  bot: Scalars['Boolean']['output'];
  /** Returns builds that this user has created. */
  builds?: Maybe<BuildConnection>;
  /** The primary email for the user */
  email: Scalars['String']['output'];
  /** Does the user have a password set */
  hasPassword: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  /** The name of the user */
  name: Scalars['String']['output'];
  /** The public UUID of the user */
  uuid: Scalars['String']['output'];
};